Improvements and simplifations following code review.
* Renamed a number of files to reduce duplication of pack in action names.
* Changed to using exceptions instead of 'sys.exit(exit_code)' for an action
  failing.
* Instead of complex nested if statements to check for keys, catch the
  KeyError and then rause an ValueError.
* Changed alias to have repo_name and packs (an array) to allow installation
  of muliple packs from a single repo.
* Stop building the repo_url in expand_repo_name.py (new name) and instead
  look up full repo URL from the config.
* Added example of a pack repo that's not in one of the StackStorm repo's.
