+++ contrib/packs/actions/deploy.yaml
@@ -1,9 +1,9 @@
 ---
-  description: "Deploy StackStorm Packs"
+  description: "Deploy StackStorm Pack(s)"
   enabled: true
-  entry_point: 'workflows/pack_deploy.yaml'
+  entry_point: 'workflows/deploy.yaml'
   
-  name: "pack_deploy"
+  name: "deploy"
   pack: "packs"
 
   parameters:
@@ -12,22 +12,22 @@
         - "expand_git_url"
         - "packs_install"
         - "deloyment_success"
+    repo_name:
+      type: "string"
+      description: "The name of the repository to expand into an Git URL."
+      required: true
     branch:
       type: "string"
       description: "The Branch to deploy."
       required: true
-    install_pack:
-      type: "string"
-      description: "The Pack to deploy."
+    packs:
+      type: "array"
+      description: "The Pack(s) to deploy."
       required: true
     auto_deploy:
       type: "boolean"
       description: "Should this pack and branch be evulated for auto deployment."
       default: false
-    key:
-      type: "string"
-      description: "The key of the commit"
-      default: ~
     message:
       type: "string"
       description: "The message for the commit"
@@ -38,7 +38,7 @@
       default: ~
 
     workflow:
-      default: packs.pack_deploy.entry
+      default: packs.deploy.entry
       immutable: true
       type: string
