+++ contrib/packs/config.yaml
@@ -1,29 +1,30 @@
 ---
-pack_group: st2packs
 
 repositories:
-  default:
-    protocol: "ssh"
-    server: "github.com"
-    port: ~
-    subtree: false
 
   # To deploy a pack from a forked st2contrib on GitHub update & uncomment the following
   # ! packs deploy me/st2contrib/<pack_name>
-  #me/st2contrib:
-  #  protocol: "https"
-  #  server: "github.com"
-  #  repo: "st2contrib"
-  #  user: "me"
-  #  port: ~
+  #<my-GitHub-user>/st2contrib:
+  #  repo: "https://github.com/<my-GitHub-user>/st2contrib.git"
   #  subtree: true
 
+  #If you have your own ST2 Pack
+  #my-st2:
+  #  repo: "https://github.com/<my-GitHub-user>/my-st2.git"
+  #  subtree: false
+  #  auto_deployment:
+  #    branch: "master"
+  #    notify_channel: "my-chatops-channel"
+
+  st2incubator:
+    repo: "https://github.com/StackStorm/st2contrib.git"
+    subtree: true
+    auto_deployment:
+      branch: "master"
+      notify_channel: "community"
+
   st2contrib:
-    protocol: "https"
-    repo: "st2contrib"
-    server: "github.com"
-    user: "StackStorm"
-    port: ~
+    repo: "https://github.com/StackStorm/st2contrib.git"
     subtree: true
     auto_deployment:
       branch: "master"
